+++ relay-general/src/store/transactions.rs
@@ -4,10 +4,20 @@ use crate::processor::{ProcessValue, ProcessingState, Processor};
 use crate::protocol::{
     Context, ContextInner, Event, EventType, Span, Timestamp, TransactionSource,
 };
-use crate::types::{Annotated, Meta, ProcessingAction, ProcessingResult};
+use crate::store::regexes::TRANSACTION_NAME_NORMALIZER_REGEX;
+use crate::types::{Annotated, Meta, ProcessingAction, ProcessingResult, Remark, RemarkType};
 
 /// Rejects transactions based on required fields.
-pub struct TransactionsProcessor;
+#[derive(Default)]
+pub struct TransactionsProcessor {
+    normalize_names: bool,
+}
+
+impl TransactionsProcessor {
+    pub fn new(normalize_names: bool) -> Self {
+        Self { normalize_names }
+    }
+}
 
 /// Get the value for a measurement, e.g. lcp -> event.measurements.lcp
 pub fn get_measurement(transaction: &Event, name: &str) -> Option<f64> {
@@ -203,6 +213,41 @@ fn set_default_transaction_source(event: &mut Event) {
     }
 }
 
+/// Normalize the transaction name.
+///
+/// Replaces UUIDs, SHAs and numerical IDs in transaction names by placeholders.
+fn normalize_transaction_name(transaction: &mut Annotated<String>) -> ProcessingResult {
+    let capture_names = TRANSACTION_NAME_NORMALIZER_REGEX
+        .capture_names()
+        .flatten()
+        .collect::<Vec<_>>();
+
+    transaction.apply(|trans, meta| {
+        // Collect all the remarks if anything matches.
+        for matches in TRANSACTION_NAME_NORMALIZER_REGEX.captures_iter(trans) {
+            for name in &capture_names {
+                if let Some(m) = matches.name(name) {
+                    let remark =
+                        Remark::with_range(RemarkType::Substituted, *name, (m.start(), m.end()));
+                    meta.add_remark(remark);
+                    break;
+                }
+            }
+        }
+
+        let changed = TRANSACTION_NAME_NORMALIZER_REGEX
+            .replace_all(trans, "*")
+            .to_string();
+        if *trans != changed && changed != "*" {
+            meta.set_original_value(Some(trans.to_string()));
+            *trans = changed
+        } else {
+            meta.clear_remarks();
+        }
+        Ok(())
+    })
+}
+
 impl Processor for TransactionsProcessor {
     fn process_event(
         &mut self,
@@ -223,6 +268,11 @@ impl Processor for TransactionsProcessor {
             event
                 .transaction
                 .set_value(Some("<unlabeled transaction>".to_owned()))
+        }
+
+        // Normalize transaction names for URLs transaction sources only.
+        if event.get_transaction_source() == &TransactionSource::Url && self.normalize_names {
+            normalize_transaction_name(&mut event.transaction)?;
         }
 
         validate_transaction(event)?;
@@ -345,7 +395,7 @@ mod tests {
         let mut event = Annotated::new(Event::default());
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -362,7 +412,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -382,7 +432,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -403,7 +453,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -425,7 +475,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -451,7 +501,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -482,7 +532,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -516,7 +566,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -554,7 +604,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -607,7 +657,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -641,7 +691,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -661,7 +711,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -716,7 +766,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -755,7 +805,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -795,7 +845,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -836,7 +886,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -878,7 +928,7 @@ mod tests {
         assert_eq!(
             process_value(
                 &mut event,
-                &mut TransactionsProcessor,
+                &mut TransactionsProcessor::default(),
                 ProcessingState::root()
             ),
             Err(ProcessingAction::InvalidTransaction(
@@ -925,7 +975,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -988,7 +1038,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -1032,7 +1082,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -1048,7 +1098,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -1096,7 +1146,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -1144,7 +1194,7 @@ mod tests {
 
         process_value(
             &mut event,
-            &mut TransactionsProcessor,
+            &mut TransactionsProcessor::default(),
             ProcessingState::root(),
         )
         .unwrap();
@@ -1227,4 +1277,220 @@ mod tests {
 
         assert!(is_high_cardinality_sdk(&event.0.unwrap()));
     }
+
+    #[test]
+    fn test_transaction_name_dont_normalize() {
+        let json = r#"
+        {
+            "type": "transaction",
+            "transaction": "/foo/2fd4e1c67a2d28fced849ee1bb76e7391b93eb12/user/123/0",
+            "transaction_info": {
+              "source": "url"
+            },
+            "timestamp": "2021-04-26T08:00:00+0100",
+            "start_timestamp": "2021-04-26T07:59:01+0100",
+            "contexts": {
+                "trace": {
+                    "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
+                    "span_id": "fa90fdead5f74053",
+                    "op": "rails.request",
+                    "status": "ok"
+                }
+            }
+        }
+        "#;
+        let mut event = Annotated::<Event>::from_json(json).unwrap();
+
+        // This must not normalize transaction name, since it's disabled.
+        process_value(
+            &mut event,
+            &mut TransactionsProcessor::default(),
+            ProcessingState::root(),
+        )
+        .unwrap();
+
+        assert_annotated_snapshot!(event, @r###"
+        {
+          "type": "transaction",
+          "transaction": "/foo/2fd4e1c67a2d28fced849ee1bb76e7391b93eb12/user/123/0",
+          "transaction_info": {
+            "source": "url"
+          },
+          "timestamp": 1619420400.0,
+          "start_timestamp": 1619420341.0,
+          "contexts": {
+            "trace": {
+              "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
+              "span_id": "fa90fdead5f74053",
+              "op": "rails.request",
+              "status": "ok",
+              "type": "trace"
+            }
+          },
+          "spans": []
+        }
+        "###);
+    }
+
+    #[test]
+    fn test_transaction_name_normalize() {
+        let json = r#"
+        {
+            "type": "transaction",
+            "transaction": "/foo/2fd4e1c67a2d28fced849ee1bb76e7391b93eb12/user/123/0",
+            "transaction_info": {
+              "source": "url"
+            },
+            "timestamp": "2021-04-26T08:00:00+0100",
+            "start_timestamp": "2021-04-26T07:59:01+0100",
+            "contexts": {
+                "trace": {
+                    "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
+                    "span_id": "fa90fdead5f74053",
+                    "op": "rails.request",
+                    "status": "ok"
+                }
+            },
+            "sdk": {"name": "sentry.ruby"},
+            "modules": {"rack": "1.2.3"}
+
+        }
+        "#;
+        let mut event = Annotated::<Event>::from_json(json).unwrap();
+
+        process_value(
+            &mut event,
+            &mut TransactionsProcessor::new(true),
+            ProcessingState::root(),
+        )
+        .unwrap();
+
+        assert_annotated_snapshot!(event, @r###"
+        {
+          "type": "transaction",
+          "transaction": "/foo/*/user/*/0",
+          "transaction_info": {
+            "source": "url"
+          },
+          "modules": {
+            "rack": "1.2.3"
+          },
+          "timestamp": 1619420400.0,
+          "start_timestamp": 1619420341.0,
+          "contexts": {
+            "trace": {
+              "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
+              "span_id": "fa90fdead5f74053",
+              "op": "rails.request",
+              "status": "ok",
+              "type": "trace"
+            }
+          },
+          "sdk": {
+            "name": "sentry.ruby"
+          },
+          "spans": [],
+          "_meta": {
+            "transaction": {
+              "": {
+                "rem": [
+                  [
+                    "sha1",
+                    "s",
+                    5,
+                    45
+                  ],
+                  [
+                    "int",
+                    "s",
+                    51,
+                    54
+                  ]
+                ],
+                "val": "/foo/2fd4e1c67a2d28fced849ee1bb76e7391b93eb12/user/123/0"
+              }
+            }
+          }
+        }
+        "###);
+    }
+
+    macro_rules! transaction_name_test {
+        ($name:ident, $input:literal, $output:literal) => {
+            #[test]
+            fn $name() {
+                let json = format!(
+                    r#"
+                    {{
+                        "type": "transaction",
+                        "transaction": "{}",
+                        "transaction_info": {{
+                          "source": "url"
+                        }},
+                        "timestamp": "2021-04-26T08:00:00+0100",
+                        "start_timestamp": "2021-04-26T07:59:01+0100",
+                        "contexts": {{
+                            "trace": {{
+                                "trace_id": "4c79f60c11214eb38604f4ae0781bfb2",
+                                "span_id": "fa90fdead5f74053",
+                                "op": "rails.request",
+                                "status": "ok"
+                            }}
+                        }}
+                    }}
+                "#,
+                    $input
+                );
+
+                let mut event = Annotated::<Event>::from_json(&json).unwrap();
+
+                process_value(
+                    &mut event,
+                    &mut TransactionsProcessor::new(true),
+                    ProcessingState::root(),
+                )
+                .unwrap();
+
+                assert_eq!($output, event.value().unwrap().transaction.value().unwrap());
+            }
+        };
+    }
+
+    transaction_name_test!(test_transaction_name_normalize_id, "/1234", "/*");
+    transaction_name_test!(
+        test_transaction_name_normalize_in_segments_1,
+        "/user/path-with-1234/",
+        "/user/*/"
+    );
+    transaction_name_test!(
+        test_transaction_name_normalize_in_segments_2,
+        "/testing/open-19-close/1",
+        "/testing/*/1"
+    );
+    transaction_name_test!(
+        test_transaction_name_normalize_sha,
+        "/hash/4c79f60c11214eb38604f4ae0781bfb2/diff",
+        "/hash/*/diff"
+    );
+    transaction_name_test!(
+        test_transaction_name_normalize_uuid,
+        "/u/7b25feea-ed2d-4132-bcbd-6232b7922add/edit",
+        "/u/*/edit"
+    );
+    transaction_name_test!(
+        test_transaction_name_normalize_hex,
+        "/u/0x3707344A4093822299F31D008/profile/123123213",
+        "/u/*/profile/*"
+    );
+    transaction_name_test!(
+        test_transaction_name_normalize_windows_path,
+        r#"C:\\\\Program Files\\1234\\Files"#,
+        r#"C:\\Program Files\*\Files"#
+    );
+    transaction_name_test!(test_transaction_name_skip_replace_all, "12345", "12345");
+    transaction_name_test!(
+        test_transaction_name_skip_replace_all2,
+        "open-12345-close",
+        "open-12345-close"
+    );
 }