+++ relay-server/src/endpoints/unreal.rs
@@ -18,7 +18,7 @@ fn extract_envelope(
     let max_payload_size = request.state().config().max_attachments_size();
 
     let future = RequestBody::new(request, max_payload_size)
-        .map_err(BadStoreRequest::PayloadError)
+        .map_err(BadStoreRequest::from)
         .and_then(move |data| {
             if data.is_empty() {
                 return Err(BadStoreRequest::EmptyBody);